Rename `getInitialRods` to `getInitialRodStates`
